Issue #SB-15614 fix: DruidEvents Validator removing duplicate code
